+++ qesource/source/caserver/SINGULAR.cpp
@@ -70,7 +70,7 @@ static Word readSingularPoly(Word r, Word V, istream &in)
   Word A, t;
   string s;
   in >> s;
-  for(unsigned int i = 0; i < s.length(); ++i)
+  for(size_t i = 0; i < s.length(); ++i)
     if (s[i] == '*') s[i] = ' ';
   s += ".\n";
   istringstream si(s);
@@ -111,7 +111,7 @@ static string WritePolyForSingular(Word r, Word P, Word V)
     out = sout.str();
   }
   // Put in * symbols
-  for(unsigned int i = 1; i < out.length() - 1; ++i)
+  for(size_t i = 1; i < out.length() - 1; ++i)
     if (out[i] == ' ' && out[i+1] != '+' && out[i+1] != '-'
 	&& out[i-1] != '+' && out[i-1] != '-'
 	)
